renaming isomorphic to business
