+++ web-client/src/useCases/createCase.js
@@ -1,5 +1,5 @@
-import Case from '../../../isomorphic/src/entities/Case';
-import User from '../../../isomorphic/src/entities/User';
+import Case from '../../../business/src/entities/Case';
+import User from '../../../business/src/entities/User';
 
 export default async (applicationContext, rawCase, rawUser) => {
   const caseDetails = new Case(rawCase);
